Update README.md
This are Bombable Carriers and Ships for Flightgear Flight Simulator and Bombable-5.0 addon. 
____________________________________________________________________________________________

To use Bombable-Carriers-Ships, install Bombabl-5.0 updated by Helijah. Install it as addon module into Flightgear.

Installation of Bombable-Carriers-Ships: Copy AI and Aircraft folder from Bombable-Carriers-Ships and paste it into your Flightgear installation under fgdata folder. You can add Bombable-Objects to custom folder, and register it in FGFS, but AI must be added into fgdata/AI. AI folder contains AI Scenarios for Bombable-Objects.

Run it as AI scenario and bring it to your location, using Bombable addon menu, or use it as "airplane" and attack it in Multiplayer session.
Spawn at waypoint over the sea or use Flightgear menu, to change its position. To operate, use trottle and joystick or set flightplan and speed in Autopilot and let it roam around. 

You can use it for landings in MP sessions only if it is static. To achieve no movement, pause Carrier. 

Remember that war is disaster for regular people and soldiers.

Aether 2025.